We don't really need double precision, so just have everything as a float

+++ ModBlobDetector.h
@@ -35,34 +35,34 @@ void ModBlobDetector::findBlobs(const cv::Mat &image, const cv::Mat &binaryImage
         Moments moms = moments(Mat(contours[contourIdx]));
         if (params.filterByArea)
         {
-            double area = moms.m00;
+            float area = moms.m00;
             if (area < params.minArea || area >= params.maxArea)
                 continue;
         }
 
         if (params.filterByCircularity)
         {
-            double area = moms.m00;
-            double perimeter = arcLength(Mat(contours[contourIdx]), true);
-            double ratio = 4 * CV_PI * area / (perimeter * perimeter);
+            float area = moms.m00;
+            float perimeter = arcLength(Mat(contours[contourIdx]), true);
+            float ratio = 4 * CV_PI * area / (perimeter * perimeter);
             if (ratio < params.minCircularity || ratio >= params.maxCircularity)
                 continue;
         }
 
         if (params.filterByInertia)
         {
-            double denominator = sqrt(pow(2 * moms.mu11, 2) + pow(moms.mu20 - moms.mu02, 2));
-            const double eps = 1e-2;
-            double ratio;
+            float denominator = sqrt(pow(2 * moms.mu11, 2) + pow(moms.mu20 - moms.mu02, 2));
+            const float eps = 1e-2;
+            float ratio;
             if (denominator > eps)
             {
-                double cosmin = (moms.mu20 - moms.mu02) / denominator;
-                double sinmin = 2 * moms.mu11 / denominator;
-                double cosmax = -cosmin;
-                double sinmax = -sinmin;
+                float cosmin = (moms.mu20 - moms.mu02) / denominator;
+                float sinmin = 2 * moms.mu11 / denominator;
+                float cosmax = -cosmin;
+                float sinmax = -sinmin;
 
-                double imin = 0.5 * (moms.mu20 + moms.mu02) - 0.5 * (moms.mu20 - moms.mu02) * cosmin - moms.mu11 * sinmin;
-                double imax = 0.5 * (moms.mu20 + moms.mu02) - 0.5 * (moms.mu20 - moms.mu02) * cosmax - moms.mu11 * sinmax;
+                float imin = 0.5 * (moms.mu20 + moms.mu02) - 0.5 * (moms.mu20 - moms.mu02) * cosmin - moms.mu11 * sinmin;
+                float imax = 0.5 * (moms.mu20 + moms.mu02) - 0.5 * (moms.mu20 - moms.mu02) * cosmax - moms.mu11 * sinmax;
                 ratio = imin / imax;
             }
             else
@@ -78,9 +78,9 @@ void ModBlobDetector::findBlobs(const cv::Mat &image, const cv::Mat &binaryImage
         {
             vector < Point > hull;
             convexHull(Mat(contours[contourIdx]), hull);
-            double area = contourArea(Mat(contours[contourIdx]));
-            double hullArea = contourArea(Mat(hull));
-            double ratio = area / hullArea;
+            float area = contourArea(Mat(contours[contourIdx]));
+            float hullArea = contourArea(Mat(hull));
+            float ratio = area / hullArea;
             if (ratio < params.minConvexity || ratio >= params.maxConvexity)
                 continue;
         }
@@ -95,13 +95,13 @@ void ModBlobDetector::findBlobs(const cv::Mat &image, const cv::Mat &binaryImage
 
         //compute blob radius
         {
-            vector<double> dists;
+            vector<float> dists;
             for (size_t pointIdx = 0; pointIdx < contours[contourIdx].size(); pointIdx++)
             {
                 Point pt = contours[contourIdx][pointIdx];
                 int xDiff = abs(center.pt.x - pt.x);
                 int yDiff = abs(center.pt.y - pt.y);
-                double dist = sqrt(pow(xDiff, 2) + pow(yDiff, 2));
+                float dist = sqrt(pow(xDiff, 2) + pow(yDiff, 2));
                 dists.push_back(dist);
             }
             std::sort(dists.begin(), dists.end());
