+++ miscutils.h
@@ -5,7 +5,7 @@
 #include <fstream>
 
 std::string toString(int i);
-std::string toString(double i);
+std::string toString(float i);
 bool fileExists(const std::string& name);
 
 #endif // MISCUTILS_H
